Bug Fix
Trying first-time experience with a sample Rails app, and realizing that
tmp/data_dumps is not created :(

Making sure this directory is created as part of initialization.
Confirming that data can be dumped locally following this change, using
a local version of the gem. Eventually a new gem version will need to be
published.